feat: save the displayName, and search users by displayName

+++ databases/user/mysql.js
@@ -1,6 +1,7 @@
-import * as properties from '../../properties/properties.js';
 import * as errors from '../../services/errors.js';
-import * as userUtils from './userUtils.js';
+import { getUserDbProperties, searchAttributes, modifiableAttributes, allAttributes } from './userUtils.js';
+import { errorIfMultiTenantContext } from '../../services/multiTenantUtils.js';
+
 import * as mysql from 'mysql2/promise';
 
 /**
@@ -9,6 +10,12 @@ import * as mysql from 'mysql2/promise';
 let connection;
 
 export async function initialize() {
+    errorIfMultiTenantContext();
+    /**
+     * @type mysql.ConnectionOptions
+     */
+    const config = getMysqlProperty();
+    config.namedPlaceholders = true;
     connection = await mysql.createConnection(getMysqlProperty());
 }
 
@@ -16,20 +23,33 @@ export function close() {
     return connection.end();
 }
 
+const selectQuery = `Select ${allAttributes.join(", ")} From ${getUserTable()} u Where u.uid = :uid`
 export async function find_user(uid) {
-    const query = "Select * From " + getUserTable() + " u Where u.uid = ?";
-    const [rows, fields] = await connection.execute(query, [uid]);
+    const [rows, fields] = await connection.execute(selectQuery, { uid: uid });
     const user = rows[0];
     return user || errors.UserNotFoundError.throw();
 }
 
+const searchQuery = `Select ${searchAttributes.join(", ")} From ${getUserTable()} Where ${searchAttributes.map(attr => `LOWER(${attr}) LIKE :token`).join(" OR ")}`;
+/**
+ * @param {String} token 
+ */
+export async function search_users(req, token) {
+    token = token.toLowerCase();
+    const [rows, fields] = await connection.execute(searchQuery, { token: `%${token}%` });
+    return rows;
+}
+
 export async function save_user(user) {
-    const selectQuery = "Select * From " + getUserTable() + " u Where u.uid = ?";
-    const [rows, fields] = await connection.execute(selectQuery, [user.uid]);
-    if (rows[0]) {
-        const updateQuery = "Update " + getUserTable() + " SET " + getMysqlProperty().transport.sms + " = ? , " + getMysqlProperty().transport.mail + " = ?  Where uid = ?";
-        const updateQueryParams = [userUtils.getSms(user), userUtils.getMail(user), user.uid];
-        await connection.execute(updateQuery, updateQueryParams);
+    const oldUser = find_user(user.uid);
+    if (oldUser) {
+        const updatedAttributes = modifiableAttributes.filter(attr => oldUser[attr] != user[attr]);
+        if (updatedAttributes.length) {
+            /** @example "sms = :sms , mail = :mail" */
+            const set = updatedAttributes.map(attr => `${attr} = :${attr}`).join(", ");
+            const updateQuery = `Update ${getUserTable()} SET ${set} Where uid = :uid`;
+            await connection.execute(updateQuery, user);
+        }
     }
 }
 
@@ -43,13 +63,13 @@ export async function create_user(uid) {
 }
 
 export function remove_user(uid) {
-    const query = "DELETE FROM " + getUserTable() + " WHERE uid=?";
-    return connection.query(query, [uid]);
+    const query = "DELETE FROM " + getUserTable() + " WHERE uid = :uid";
+    return connection.query(query, { uid: uid });
 }
 
 
 function getMysqlProperty() {
-    return properties.getEsupProperty('mysql');
+    return getUserDbProperties();
 }
 
 function getUserTable() {
